Adds more global location lookup
This switches to one central source for location information. This is
downloaded using the Makefile in CSV_database and removes the need for
separate databases per country. The generation of results is modified as
we now load the entire dataset before searching, however this adds only
a few seconds onto the search time. The country codes are now defined by
the downloaded database.