Adds automation of results collections
This adds a script CROBJOB which can be added to crontab to auomate the
process of downloading data. A country is picked at random from those
with N > 1000 number of data points.

Added functionality is given to basic_tools.py to enable both the random
country argument, and the check of the number of data points.